+++ buildSrc/src/main/java/org/elasticsearch/gradle/test/rest/transform/headers/InjectHeaders.java
@@ -8,26 +8,21 @@
 
 package org.elasticsearch.gradle.test.rest.transform.headers;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.databind.node.TextNode;
 import org.elasticsearch.gradle.test.rest.transform.RestTestTransform;
 import org.elasticsearch.gradle.test.rest.transform.RestTestTransformByParentObject;
-import org.elasticsearch.gradle.test.rest.transform.RestTestTransformGlobalSetup;
-import org.elasticsearch.gradle.test.rest.transform.RestTestTransformGlobalTeardown;
+import org.elasticsearch.gradle.test.rest.transform.feature.FeatureInjector;
 import org.gradle.api.tasks.Input;
 
-import javax.annotation.Nullable;
-import java.util.Iterator;
 import java.util.Map;
 
 /**
  * A {@link RestTestTransform} that injects HTTP headers into a REST test. This includes adding the necessary values to the "do" section
  * as well as adding headers as a features to the "setup" and "teardown" sections.
  */
-public class InjectHeaders implements RestTestTransformByParentObject, RestTestTransformGlobalSetup, RestTestTransformGlobalTeardown {
+public class InjectHeaders extends FeatureInjector implements RestTestTransformByParentObject {
 
     private static JsonNodeFactory jsonNodeFactory = JsonNodeFactory.withExactBigDecimals(false);
 
@@ -59,98 +54,8 @@ public class InjectHeaders implements RestTestTransformByParentObject, RestTestT
     }
 
     @Override
-    public ObjectNode transformSetup(ObjectNode setupNodeParent) {
-        // check to ensure that headers feature does not already exist
-        if (setupNodeParent != null) {
-            ArrayNode setupNode = (ArrayNode) setupNodeParent.get("setup");
-            if (hasHeadersFeature(setupNode)) {
-                return setupNodeParent;
-            }
-        }
-        // transform or insert the headers feature into setup/skip/features
-        ArrayNode setupNode;
-        if (setupNodeParent == null) {
-            setupNodeParent = new ObjectNode(jsonNodeFactory);
-            setupNode = new ArrayNode(jsonNodeFactory);
-            setupNodeParent.set("setup", setupNode);
-        }
-        setupNode = (ArrayNode) setupNodeParent.get("setup");
-        addSkip(setupNode);
-        return setupNodeParent;
-    }
-
-    @Override
-    public ObjectNode transformTeardown(@Nullable ObjectNode teardownNodeParent) {
-        if (teardownNodeParent != null) {
-            ArrayNode teardownNode = (ArrayNode) teardownNodeParent.get("teardown");
-            // only transform an existing teardown section since a teardown does not inherit from setup but still needs the skip section
-            if (teardownNode != null) {
-                // check to ensure that headers feature does not already exist
-                if (hasHeadersFeature(teardownNode)) {
-                    return teardownNodeParent;
-                }
-                addSkip(teardownNode);
-                return teardownNodeParent;
-            }
-        }
-        return teardownNodeParent;
-    }
-
-    private boolean hasHeadersFeature(ArrayNode skipParent) {
-        JsonNode features = skipParent.at("/0/skip/features");
-        if (features != null) {
-            if (features.isArray()) {
-                ArrayNode featuresArray = (ArrayNode) features;
-                Iterator<JsonNode> it = featuresArray.elements();
-                while (it.hasNext()) {
-                    if ("headers".equals(it.next().asText())) {
-                        return true;
-                    }
-                }
-            } else {
-                if ("headers".equals(features.asText())) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-
-    private void addSkip(ArrayNode skipParent) {
-        Iterator<JsonNode> skipParentIt = skipParent.elements();
-        boolean foundSkipNode = false;
-        while (skipParentIt.hasNext()) {
-            JsonNode arrayEntry = skipParentIt.next();
-            if (arrayEntry.isObject()) {
-                ObjectNode skipCandidate = (ObjectNode) arrayEntry;
-                if (skipCandidate.get("skip") != null) {
-                    ObjectNode skipNode = (ObjectNode) skipCandidate.get("skip");
-                    foundSkipNode = true;
-                    JsonNode featuresNode = skipNode.get("features");
-                    if (featuresNode == null) {
-                        ObjectNode featuresNodeObject = new ObjectNode(jsonNodeFactory);
-                        skipNode.set("features", TextNode.valueOf("headers"));
-                    } else if (featuresNode.isArray()) {
-                        ArrayNode featuresNodeArray = (ArrayNode) featuresNode;
-                        featuresNodeArray.add("headers");
-                    } else if (featuresNode.isTextual()) {
-                        // convert to an array
-                        ArrayNode featuresNodeArray = new ArrayNode(jsonNodeFactory);
-                        featuresNodeArray.add(featuresNode.asText());
-                        featuresNodeArray.add("headers");
-                        // overwrite the features object
-                        skipNode.set("features", featuresNodeArray);
-                    }
-                }
-            }
-        }
-        if (foundSkipNode == false) {
-            ObjectNode skipNode = new ObjectNode(jsonNodeFactory);
-            ObjectNode featuresNode = new ObjectNode(jsonNodeFactory);
-            skipParent.insert(0, skipNode);
-            featuresNode.set("features", TextNode.valueOf("headers"));
-            skipNode.set("skip", featuresNode);
-        }
+    public String getSkipFeatureName() {
+        return "headers";
     }
 
     @Input
