+++ buildSrc/src/test/resources/rest/transform/feature/with_multiple_feature.yml
@@ -1,0 +1,13 @@
+---
+setup:
+  - skip:
+      features:
+      - pre_existing_feature1
+      - pre_existing_feature2
+---
+"Test without with multiple feature setup":
+  - do:
+      something:
+        id: "something"
+  - match: { acknowledged: true }
+
