+++ buildSrc/src/test/resources/rest/transform/feature/with_setup_no_skip.yml
@@ -1,0 +1,13 @@
+---
+setup:
+  - do:
+      some.setup:
+        index: blah
+---
+"Test without with setup but no skip (and by inference no features)":
+  - do:
+      something:
+        id: "something"
+  - match: { acknowledged: true }
+
+
