+++ buildSrc/src/test/resources/rest/transform/feature/with_single_feature.yml
@@ -1,0 +1,11 @@
+---
+setup:
+  - skip:
+      features: pre_existing_feature
+---
+"Test without with single feature setup":
+  - do:
+      something:
+        id: "something"
+  - match: { acknowledged: true }
+
