+++ buildSrc/src/test/resources/rest/transform/header/with_existing_headers.yml
@@ -1,0 +1,9 @@
+---
+"Test without a setup":
+  - do:
+      headers:
+        foo: "bar"
+      something:
+        id: "something"
+  - match: { acknowledged: true }
+
