--- buildSrc/src/test/resources/rest/transform/header/with_features.yml
@@ -1,25 +1,0 @@
----
-setup:
-  - skip:
-      features: allowed_warnings
-
-  - do:
-      cluster.health:
-        wait_for_status: yellow
-
----
-"Test inline template against specific index":
-  - do:
-      search_template:
-        rest_total_hits_as_int: true
-        index: foobar
-        body:
-          source:
-            query:
-              match:
-                title: "{{query_string}}"
-          params:
-            query_string: "search for these words"
-
-  - match: { hits.total: 1}
-  - match: { hits.hits.0._id: "1"}
