+++ buildSrc/src/test/resources/rest/transform/header/without_existing_headers.yml
@@ -1,0 +1,7 @@
+---
+"Test without a setup":
+  - do:
+      something:
+        id: "something"
+  - match: { acknowledged: true }
+
